Laget pattern for oversetting og zod skjema for ArbeidstakerenSteg
