+++ app/src/pages/skjema/arbeidstaker/ArbeidstakerenSteg.tsx
@@ -2,6 +2,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { Select, TextField, VStack } from "@navikt/ds-react";
 import { useNavigate } from "@tanstack/react-router";
 import { FormProvider, useForm } from "react-hook-form";
+import { useTranslation } from "react-i18next";
 import { z } from "zod";
 
 import { DatePickerFormPart } from "~/components/DatePickerFormPart.tsx";
@@ -14,145 +15,20 @@ import {
   SkjemaSteg,
 } from "~/pages/skjema/components/SkjemaSteg.tsx";
 
+import {
+  AKTIVITET_OPTIONS,
+  arbeidstakerSchema,
+} from "./arbeidstakerenStegSchema.ts";
+
 const stepKey = "arbeidstakeren";
-
-const AKTIVITET_OPTIONS = [
-  { value: "studier", label: "Studier" },
-  { value: "ferie", label: "Ferie" },
-  { value: "selvstendig-virksomhet", label: "Selvstendig virksomhet" },
-  { value: "kontantytelse-fra-nav", label: "Mottok kontantytelse fra Nav" },
-] as const;
-
-const baseArbeidstakerSchema = z.object({
-  harVaertEllerSkalVaereILonnetArbeidFoerUtsending: z.boolean({
-    message:
-      "Du må svare på om du har vært eller skal være i lønnet arbeid i Norge før utsending",
-  }),
-  aktivitetIMaanedenFoerUtsendingen: z.string().optional(),
-  skalJobbeForFlereVirksomheter: z.boolean({
-    message:
-      "Du må svare på om du skal jobbe for flere virksomheter i perioden",
-  }),
-  norskeVirksomheterArbeidstakerJobberForIutsendelsesPeriode: z
-    .array(
-      z.object({
-        organisasjonsnummer: z
-          .string()
-          .min(1, "Organisasjonsnummer er påkrevd")
-          .regex(/^\d{9}$/, "Organisasjonsnummer må være 9 siffer"),
-      }),
-    )
-    .optional(),
-  utenlandskeVirksomheterArbeidstakerJobberForIutsendelsesPeriode: z
-    .array(
-      z.object({
-        navn: z.string().min(1, "Navn på virksomhet er påkrevd"),
-        organisasjonsnummer: z.string().optional(),
-        vegnavnOgHusnummer: z
-          .string()
-          .min(1, "Vegnavn og husnummer er påkrevd"),
-        bygning: z.string().optional(),
-        postkode: z.string().optional(),
-        byStedsnavn: z.string().optional(),
-        region: z.string().optional(),
-        land: z.string().min(1, "Land er påkrevd"),
-        tilhorerSammeKonsern: z.boolean({
-          message: "Du må svare på om virksomheten tilhører samme konsern",
-        }),
-      }),
-    )
-    .optional(),
-  harNorskFodselsnummer: z.boolean({
-    message: "Du må svare på om arbeidstakeren har norsk fødselsnummer",
-  }),
-  fodselsnummer: z.string().optional(),
-  fornavn: z.string().optional(),
-  etternavn: z.string().optional(),
-  fodselsdato: z.string().optional(),
-});
-
-type BaseArbeidstakerFormData = z.infer<typeof baseArbeidstakerSchema>;
-
-function validerFodselsdato(data: BaseArbeidstakerFormData) {
-  if (!data.harNorskFodselsnummer) {
-    return data.fodselsdato && data.fodselsdato.length > 0;
-  }
-  return true;
-}
-
-function validerEtternavn(data: BaseArbeidstakerFormData) {
-  if (!data.harNorskFodselsnummer) {
-    return data.etternavn && data.etternavn.length >= 2;
-  }
-  return true;
-}
-
-function validerFodselsnummer(data: BaseArbeidstakerFormData) {
-  if (data.harNorskFodselsnummer) {
-    return data.fodselsnummer && data.fodselsnummer.length > 0;
-  }
-  return true;
-}
-
-function validerFodselsnummerFormat(data: BaseArbeidstakerFormData) {
-  if (data.fodselsnummer && data.fodselsnummer.length > 0) {
-    return /^\d{11}$/.test(data.fodselsnummer);
-  }
-  return true;
-}
-
-function validerFornavn(data: BaseArbeidstakerFormData) {
-  if (!data.harNorskFodselsnummer) {
-    return data.fornavn && data.fornavn.length >= 2;
-  }
-  return true;
-}
-
-function validerHarVaertEllerSkalVaereILonnetArbeidFoerUtsending(
-  data: BaseArbeidstakerFormData,
-) {
-  if (data.harVaertEllerSkalVaereILonnetArbeidFoerUtsending) {
-    return true;
-  }
-
-  const validOptions = AKTIVITET_OPTIONS.map((opt) => opt.value);
-  return (
-    data.aktivitetIMaanedenFoerUtsendingen &&
-    (validOptions as string[]).includes(data.aktivitetIMaanedenFoerUtsendingen)
-  );
-}
-
-const arbeidstakerSchema = baseArbeidstakerSchema
-  .refine(validerFodselsnummer, {
-    message:
-      "Fødselsnummer eller d-nummer er påkrevd når arbeidstakeren har norsk fødselsnummer",
-    path: ["fodselsnummer"],
-  })
-  .refine(validerFodselsnummerFormat, {
-    message: "Fødselsnummer eller d-nummer må være 11 siffer",
-    path: ["fodselsnummer"],
-  })
-  .refine(validerFornavn, {
-    message: "Fornavn er påkrevd og må være minst 2 tegn",
-    path: ["fornavn"],
-  })
-  .refine(validerEtternavn, {
-    message: "Etternavn er påkrevd og må være minst 2 tegn",
-    path: ["etternavn"],
-  })
-  .refine(validerFodselsdato, {
-    message: "Fødselsdato er påkrevd",
-    path: ["fodselsdato"],
-  })
-  .refine(validerHarVaertEllerSkalVaereILonnetArbeidFoerUtsending, {
-    message: "Du må velge en aktivitet når du ikke har vært i lønnet arbeid",
-    path: ["aktivitetIMaanedenFoerUtsendingen"],
-  });
-
-type ArbeidstakerFormData = z.infer<typeof arbeidstakerSchema>;
 
 export function ArbeidstakerenSteg() {
   const navigate = useNavigate();
+  const { t } = useTranslation();
+
+  const translateError = (error?: string) => error && t(error);
+
+  type ArbeidstakerFormData = z.infer<typeof arbeidstakerSchema>;
 
   const formMethods = useForm<ArbeidstakerFormData>({
     resolver: zodResolver(arbeidstakerSchema),
@@ -187,21 +63,24 @@ export function ArbeidstakerenSteg() {
         <SkjemaSteg
           config={{
             stepKey,
-            customNesteKnapp: { tekst: "Lagre og fortsett", type: "submit" },
+            customNesteKnapp: {
+              tekst: t("felles.lagreOgFortsett"),
+              type: "submit",
+            },
             stegRekkefolge: ARBEIDSTAKER_STEG_REKKEFOLGE,
           }}
         >
           <RadioGroupJaNeiFormPart
             className="mt-4"
             formFieldName="harNorskFodselsnummer"
-            legend="Har arbeidstakeren norsk fødselsnummer eller d-nummer?"
+            legend={t("arbeidstaker.harNorskFodselsnummer")}
           />
 
           {harNorskFodselsnummer && (
             <TextField
               className="mt-4"
-              error={errors.fodselsnummer?.message}
-              label="Arbeidstakerens fødselsnummer eller d-nummer"
+              error={translateError(errors.fodselsnummer?.message)}
+              label={t("arbeidstaker.fodselsnummerEllerDNummer")}
               size="medium"
               style={{ maxWidth: "160px" }}
               {...register("fodselsnummer")}
@@ -212,22 +91,22 @@ export function ArbeidstakerenSteg() {
             <>
               <TextField
                 className="mt-4 max-w-md"
-                error={errors.fornavn?.message}
-                label="Arbeidstakerens fornavn"
+                error={translateError(errors.fornavn?.message)}
+                label={t("arbeidstaker.fornavn")}
                 {...register("fornavn")}
               />
 
               <TextField
                 className="mt-4 max-w-md"
-                error={errors.etternavn?.message}
-                label="Arbeidstakerens etternavn"
+                error={translateError(errors.etternavn?.message)}
+                label={t("arbeidstaker.etternavn")}
                 {...register("etternavn")}
               />
 
               <DatePickerFormPart
                 className="mt-4"
                 formFieldName="fodselsdato"
-                label="Arbeidstakerens fødselsdato"
+                label={t("arbeidstaker.fodselsdato")}
               />
             </>
           )}
@@ -235,24 +114,26 @@ export function ArbeidstakerenSteg() {
           <RadioGroupJaNeiFormPart
             className="mt-4"
             formFieldName="harVaertEllerSkalVaereILonnetArbeidFoerUtsending"
-            legend="Har du vært eller skal du være i lønnet arbeid i Norge i minst én måned rett før utsendingen?"
+            legend={t("arbeidstaker.harVaertEllerSkalVaereILonnetArbeid")}
           />
 
           {harVaertEllerSkalVaereILonnetArbeidFoerUtsending === false && (
             <Select
               className="mt-4"
-              error={errors.aktivitetIMaanedenFoerUtsendingen?.message}
-              label="Aktivitet (TODO: finne en bra label her)"
+              error={translateError(
+                errors.aktivitetIMaanedenFoerUtsendingen?.message,
+              )}
+              label={t("arbeidstaker.aktivitet")}
               style={{ width: "fit-content" }}
               {...register("aktivitetIMaanedenFoerUtsendingen")}
             >
-              <option value="">Velg aktivitet</option>
+              <option value="">{t("arbeidstaker.velgAktivitet")}</option>
               {AKTIVITET_OPTIONS.map((aktivitetOption) => (
                 <option
                   key={aktivitetOption.value}
                   value={aktivitetOption.value}
                 >
-                  {aktivitetOption.label}
+                  {t(aktivitetOption.labelKey)}
                 </option>
               ))}
             </Select>
@@ -261,7 +142,7 @@ export function ArbeidstakerenSteg() {
           <RadioGroupJaNeiFormPart
             className="mt-4"
             formFieldName="skalJobbeForFlereVirksomheter"
-            legend="Skal du jobbe for flere virksomheter i perioden du søker for?"
+            legend={t("arbeidstaker.skalJobbeForFlereVirksomheter")}
           />
 
           {skalJobbeForFlereVirksomheter === true && (
